Refactor: move EditButton component for consistency
Internal to table component, unrelated to editor
