+++ packages/admin-panel/src/editor/index.js
@@ -6,7 +6,6 @@
 export { CreateButton } from './CreateButton';
 export { DATA_CHANGE_ACTIONS } from './constants';
 export { DataChangeAction } from './DataChangeAction';
-export { EditButton } from './EditButton';
 export { BulkEditButton } from './BulkEditButton';
 export { BulkCreateButton } from './BulkCreateButton';
 export { EditModal } from './EditModal';
