+++ packages/admin-panel/src/table/columnTypes/generateConfigForColumnType.js
@@ -4,13 +4,14 @@
  */
 
 import React from 'react';
-import { EditButton, BulkEditButton } from '../../editor';
+import { BulkEditButton } from '../../editor';
 import { DeleteButton } from './DeleteButton';
 import { ExportButton } from '../../importExport';
 import { BooleanSelectFilter } from './columnFilters';
 import { Tooltip, JSONTooltip } from './Tooltip';
 import { LogsButton } from '../../logsTable';
 import { SyncStatus } from '../../sync';
+import { EditButton } from './EditButton';
 
 const generateCustomCell = (CustomCell, actionConfig, reduxId) => props => (
   <CustomCell actionConfig={actionConfig} reduxId={reduxId} {...props} />
